+++ src/main/kotlin/com/zepben/evolve/services/network/NetworkService.kt
@@ -29,7 +29,7 @@ import com.zepben.evolve.cim.iec61970.base.wires.generation.production.PowerElec
 import com.zepben.evolve.cim.iec61970.infiec61970.feeder.Circuit
 import com.zepben.evolve.cim.iec61970.infiec61970.feeder.Loop
 import com.zepben.evolve.services.common.BaseService
-import com.zepben.evolve.services.network.tracing.ConnectivityResult
+import com.zepben.evolve.services.network.tracing.connectivity.ConnectivityResult
 import com.zepben.evolve.services.network.tracing.phases.NominalPhasePath
 import com.zepben.evolve.services.network.tracing.phases.XyPhaseStep
 import java.util.*